refactor: clean up schemas, public API, and BEATs warning
- Replace dtos/ with schemas/ (types.py, items.py, predictions.py) to avoid circular imports
- Remove internal Constructor names and registry dicts from __all__
- Fix _apply_segmentation to build a new list instead of mutating self.items
- Suppress FutureWarning from deprecated weight_norm in BEATs backbone
- Add publish.yml GitHub Actions workflow for automated PyPI releases
- Update tests to use public API aliases (AudioClassifier, Backbone)
