+++ src/deepaudiox/schemas/types.py
@@ -1,0 +1,4 @@
+from typing import Literal
+
+BackboneName = Literal["beats", "passt", "mobilenet_05_as", "mobilenet_10_as", "mobilenet_40_as"]
+"""Supported pretrained backbone names."""
